% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ass}
\alias{ass}
\title{Generates all possible assignments of survey participants.}
\usage{
ass(n)
}
\arguments{
\item{n}{- single integer - number of participants}
}
\value{
- matrix
}
\description{
This function generates all possible 0-1 assignment
vectors of 2n survey participants in such a way that
exactly n of them are assigned to group 0 (control)
and the other n ones are assigned to group 1 (treatment).
}

