% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IPIPANRcourseHomework-package.R
\docType{package}
\name{IPIPANRcourseHomework-package}
\alias{IPIPANRcourseHomework-package}
\title{R course homework}
\description{
This package contains solutions for 6th homework
of IPIPAN Advanced R course.
}

