% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{Returns vector's mode.}
\usage{
mode(x1)
}
\arguments{
\item{x1}{- integer vector}
}
\value{
- integer
}
\description{
This function calculates the most frequently
occurring value in an integer vector (mode).
If the mode is ambiguous returns the lowest mode.
Please note that "mode" overrides mode function
from base package.
}

