% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/homework6_1.R
\name{mode2}
\alias{mode2}
\title{Returns vector's mode.}
\usage{
mode2(x)
}
\arguments{
\item{x}{- integer vector}
}
\value{
- integer
}
\description{
This function calculates the most frequently
occurring value in an integer vector (mode).
If the mode is ambiguous returns the lowest mode.
This is R reference function for benchmark tests.
}

